migrated all subspaces-related test to test_subspaces.py
